refactor(config): config struct refactor.
